Add RawQuery and Post methods
The RawQuery method allows us to pass in a query with all the URL
processing and authentication handled, but the output received as an
io.ReadCloser instead.  The specific use for this was to grab a config
file.

The Post method is to allow us to post things up while leveraging all
the rsn authentication information that a *Server has.